Add InsertRecords to SQL (batch)

+++ src/main/java/net/usachev/magnit/converter/SAXSumElements.java
@@ -6,16 +6,16 @@ import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
 public class SAXSumElements extends DefaultHandler {
-    private int sum = 0;
+    private long sum = 0;
 
     @Override
     public void startElement(String uri, String localName, String qName, Attributes attributes) throws SAXException {
         if (qName.equals("entry")) {
-            sum += Integer.valueOf(attributes.getValue(0));
+            sum += Long.valueOf(attributes.getValue(0));
         }
     }
 
-    public int getSum() {
+    public long getSum() {
         return sum;
     }
 }
